requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
faster-whisper
numpy
